classes moved to their own folders
- I managed to run the commands to see if there were any typechecks going wrong
- added the class launcher to the main.py and removed it from the directory
- removed the import from CLI game for debug reasons
- added mypy dependency in requirements_dev.txt
- as I said in an old PR the files that run the gameloop are still called gameloop.py, in it's own directory/folder
- inventory class had a docstring with apend, is now append
