+++ requirements_dev.txt
@@ -1,1 +1,2 @@
 ruff==0.5.3
+mypy==1.11.1
